move handlers into their own file and clean up poutine namespace
